+++ lib/idx/mixinMinimal.ts
